server: extract debug from admin
